% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_mean}
\alias{day_mean}
\title{Calculates the mean of variables at the day level}
\usage{
day_mean(data, variables, grouping = c("person_id", "daynr"))
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{grouping}{A character vector of grouping variables at which you want to calculate the mean. This should be your person id variable followed by your day variable.
defaults to c('person_id', 'daynr')}
}
\description{
This is a simple function that returns a dataframe in which the day-level mean of any input variables
has been calculated. Output variables will be named 'input_variable_day_mean'
}
