% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_min}
\alias{day_min}
\title{Calculates the min of variables at the day level}
\usage{
day_min(data, variables, person_id, day_id)
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{person_id}{The person level grouping variable - this should be your person id variable.}

\item{day_id}{The day level grouping variable - this should be your day id variable.}
}
\description{
This is a simple function that returns a dataframe in which the day-level min of any input variables
has been calculated. Output variables will be named 'input_variable_day_min'
}
