% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_min_sd}
\alias{day_min_sd}
\title{Calculates the (person level) standard deviation of daily minima of variables}
\usage{
day_min_sd(data, variables, grouping = c("person_id", "daynr"))
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{grouping}{The grouping variables at which you want to calculate the mean. This should be your person id variable 
followed by your day identifier. defaults to c('person_id', 'daynr')}
}
\description{
This is a function that returns a dataframe in which the person-level standard deviation of the daily minimums
of any input variables has been calculated. Output variables will be named 'input_variable_day_min_sd'.
Note carefully that this is -not- the min of the standard deviations within each day. Rather, the minimum
at each day is taken, then the standard deviation of these minima is calculated.
}
