% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_sd}
\alias{day_sd}
\title{Calculates the sd of variables at the day level}
\usage{
day_sd(data, vars, grouping = c("person_id", "daynr"))
}
\arguments{
\item{data}{The input dataset}

\item{vars}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{grouping}{A character vector of grouping variables at which you want to calculate the sd. This should be your person id variable followed by your day variable.
defaults to c('person_id', 'daynr')}
}
\description{
This is a simple function that returns a dataframe in which the day-level sd of any input variables
has been calculated. Output variables will be named 'input_variable_day_sd'
}
