% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{esm_lag}
\alias{esm_lag}
\title{Lag observation-level variables}
\usage{
esm_lag(data, vars, grouping = c("person_id", "daynr"), order = "obs_id")
}
\arguments{
\item{data}{The dataset to operate on.}

\item{vars}{The character vector of variables to lag}

\item{grouping}{The character vector of grouping variables (by default, c('person_id', 'daynr'))}

\item{order}{The observation number to order by within days. By default it is 'obs_id'. You will want
this to be a unique, ordered observation identifier. It does not matter whether it is unique only within
days (i.e. resets each day for each p).}
}
\description{
This function will lag any variable on the observation level. It takes a dataset, a character vector of variables, 
and a character vector of grouping variables (by default the person and day ids)
and lags within these groups. With the default grouping levels, this will allow lags within day, but will not allow lags across day (the first lagged
observation of each day for each P will always be NA).
}
\details{
It can't be used for lagging day level variables because those need to be aggregated or sliced to one obs per day first - the dplyr 
lag function this is built on doesn't lag at the grouping level, it just lags within groups by 1 row.

\code{\link{esm_lag}} and \code{\link{esm_day_lag}} are unique in that they output variables with the same appelation (_lag) because they output what we would expect
for lag at each level so this intuitively makes sense. The variables will already have names to say whether they're at the day level or not.
}
