% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keep_bottom}
\alias{keep_bottom}
\title{Keeps only variables that -do- vary within the grouping variables}
\usage{
keep_bottom(data, grouping)
}
\arguments{
\item{data}{The input dataset}

\item{grouping}{The grouping variables that determine the level at which variables should not vary if
they are to be kept.}
}
\description{
Effectively, this function keeps variables that are -not- invariant below the grouping level specified. For
example, if person_id was specified, it would keep only variables that do vary below the person
level. It does this by calling keep_top and excluding anything that does not vary below the person level
(i.e. trait level variables). If person_id and daynr were specified, it would keep only variables
that vary below the day level (i.e. momentary variables). Grouping variables are always retained.
}
