% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keep_top}
\alias{keep_top}
\title{Keeps only variables that do not vary below the level of the grouping variables}
\usage{
keep_top(data, grouping)
}
\arguments{
\item{data}{The input dataset}

\item{grouping}{The grouping variables that determine the level below which variables should not vary if
they are to be kept.}
}
\description{
Effectively, this function keeps variables that are 'at the level of' the grouping level specified. For
example, if person_id was specified, it would keep only variables that do not vary below the person
level (i.e. trait level variables). Grouping variables are always retained. If person_id and daynr were specified, it would keep only variables
that do not vary within day (i.e. day level variables).
}
