% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_rescale}
\alias{my_rescale}
\title{A utility function that makes rescale_cols work. Usually not useful directly.}
\usage{
my_rescale(x, new.min, new.max, old.min, old.max)
}
\description{
A utility function that makes rescale_cols work. Usually not useful directly.
}
