% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{operate_cols}
\alias{operate_cols}
\title{Flexibly operate on sets of variables within a dataframe}
\usage{
operate_cols(data, variables, operation = "add", number)
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of variables to be operated on}

\item{operation}{one of 'add', 'subtract', 'multiply', 'divide'}

\item{number}{A constant. This should usually be positive (subtract will automatically flip it), and will
make sense to say out loud "subtract 2" rather than "subtract -2" is usually what is intended. Of course,
you can multiply by -4, if for some reason you so wish...}
}
\description{
This simple function allows for simple operations to be performed on sets of variables within a dataframe.
It is useful mainly for maintaining code consistency and saving subsetting code. It can add, divide, 
multiply and subtract by a constant, and will perform the same operation on all variables passed in.
}
\examples{
# Subtract one from the carb column of mtcars
operate_cols(mtcars, 'carb', 'subtract', 1)

}
