% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{person_max}
\alias{person_max}
\title{Calculates the max of variables at the person level}
\usage{
person_max(data, variables, grouping = c("person_id"))
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{grouping}{The grouping variables at which you want to calculate the mean. This should be your person id variable defaults to 'person_id'}
}
\description{
This is a simple function that returns a dataframe in which the person-level max of any input variables
has been calculated. Output variables will be named 'input_variable_max'
}
