% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{person_mean}
\alias{person_mean}
\title{Calculates the mean of variables at the person level}
\usage{
person_mean(data, variables, person_id)
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{person_id}{The grouping variable to calculate your mean at - this should be your person id variable.}
}
\description{
This is a simple function that returns a dataframe in which the person-level mean of any input variables
has been calculated. Output variables will be named 'input_variable_mean'
}
\details{
Technically, it will calculate the mean at any level, but I have used simple dplyr
to set the output to 'original_variable_name_mean'. It is technically possible to write functions that
flexible assign the output name so a simple 'esm_mean' function could be created to work at any level and
allow the user to specify output variable names. But I made a design decision to be moderately constrictive
here for ease of use and also to ensure that variables are consistently named coming out of these functions.
}
