% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename a series of columns in a dataframe}
\usage{
rename_cols(data, ...)
}
\arguments{
\item{data}{The input dataset}

\item{...}{A series of key value pairs (in character form) of target variables and what they currently are in the dataset}
}
\description{
Given a dataframe and a series of key-value pairs, replaced all the key matches within the column names of the dataframe
with their corresponding values, and returns a data frame as a result.
}
\examples{
# Turns the Girth column in trees into an Alpha column
rename_cols(trees, 'Alpha' = 'Girth')

}
