% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rescale_cols}
\alias{rescale_cols}
\title{A simple (but very useful) function to rescale variables flexibly.}
\usage{
rescale_cols(data, variables, old.min, old.max, new.min, new.max)
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of variables to be operated on}

\item{old.min}{the minima of the current variables}

\item{old.max}{the maxima of the current variables}

\item{new.min}{the minima of the new variables}

\item{new.max}{the maxima of the new variables}
}
\description{
This function allows for a set of variables to be rescaled from any original min and max to any target
min and max.
}
\examples{
# rescale the carb column of mtcars from 1-8 to 1-100.
rescale_cols(mtcars, 'carb', old.min = 1, old.max = 8, new.min = 1, new.max = 100)

}
