% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reverse_cols}
\alias{reverse_cols}
\title{Reverse a series of columns in a dataframe}
\usage{
reverse_cols(data, variables, min, max)
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of variables that require reversing}

\item{min}{The minimum of all the variables to be reversed (for variables that have different mins, use separate
function calls)}

\item{max}{The maximum of all the variables to be reversed}
}
\description{
Given a dataframe, a list of variables, and the min and max of those variables (which must be the same)
reverses all variables. Returns a dataframe containing the result.
}
\examples{
# Reverses the carb column in mtcars
reverse_cols(mtcars, 'carb', min = 1, max = 8)

}
