% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trim_min_obs}
\alias{trim_min_obs}
\title{Keep only groups (usually days) with enough observations.}
\usage{
trim_min_obs(data, grouping, min.obs = 2)
}
\arguments{
\item{data}{The input dataset}

\item{grouping}{The grouping variables within which you require a minimum observation number. Could be just
your person id or your person id and day. Should take the form of a character vector.}

\item{min.obs}{The minimum number of observations to require. Defaults to 2.}
}
\description{
This function allows for trimming esm data down so that only days (or people) with a minimum number of observations
are included. If there is a 'missed' variable (coded 1 = missed) these rows will not be counted. To 
make sure enough observations of a particular variable are included, use \code{trim_min_valid_obs}.
}
\examples{
# Trim the sleep dataset to have 2 obs per ID (this does nothing, but illustrates syntax)
trim_min_obs(sleep, grouping = 'ID', min.obs = 2)

}
