% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_max_relsd}
\alias{day_max_relsd}
\title{Calculates the (person level) relative standard deviation of daily maxima of variables}
\usage{
day_max_relsd(data, vars, min, max, grouping = c("person_id", "daynr"))
}
\arguments{
\item{data}{The input dataset}

\item{vars}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{grouping}{The grouping variables at which you want to calculate the mean. This should be your person id variable 
followed by your day identifier. defaults to c('person_id', 'daynr')}
}
\description{
This is a function that returns a dataframe in which the person-level relative standard deviation of the daily maximums
of any input variables has been calculated. Output variables will be named 'input_variable_day_max_relsd'.
Note carefully that this is -not- the max of the relative standard deviations within each day. Rather, the maximum
at each day is taken, then the relative standard deviation of these maxima is calculated.
}
