% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_mean}
\alias{day_mean}
\title{Calculates the mean of variables at the day level}
\usage{
day_mean(data, variables, person_id, day_id)
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{person_id}{The person level grouping variable - this should be your person id variable.}

\item{day_id}{The day level grouping variable - this should be your day id variable.}
}
\description{
This is a simple function that returns a dataframe in which the day-level mean of any input variables
has been calculated. Output variables will be named 'input_variable_day_mean'
}
