% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_min}
\alias{day_min}
\title{Calculates the min of variables at the day level}
\usage{
day_min(data, variables, grouping = c("person_id", "daynr"))
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{grouping}{A character vector of grouping variables at which you want to calculate the min. This should be your person id variable followed by your day variable.
defaults to c('person_id', 'daynr')}
}
\description{
This is a simple function that returns a dataframe in which the day-level min of any input variables
has been calculated. Output variables will be named 'input_variable_day_min'
}
