% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{person_min}
\alias{person_min}
\title{Calculates the min of variables at the person level}
\usage{
person_min(data, variables, person_id)
}
\arguments{
\item{data}{The input dataset}

\item{variables}{A character vector of input variables e.g. c('negemo', 'posemo').}

\item{person_id}{The grouping variable to calculate your mean at - this should be your person id variable.}
}
\description{
This is a simple function that returns a dataframe in which the person-level min of any input variables
has been calculated. Output variables will be named 'input_variable_max'
}
